Move MetricalDurationSpanningContainer to own file
